@@ -4,13 +4,16 @@
 	"description": "Control the macOS dark mode",
 	"license": "MIT",
 	"repository": "sindresorhus/node-dark-mode",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=4"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -23,13 +26,15 @@
 		"dark",
 		"mode",
 		"control",
-		"toggle"
+		"toggle",
+		"automate",
+		"automation"
 	],
 	"dependencies": {
-		"run-jxa": "^1.0.0"
+		"run-jxa": "^3.0.0"
 	},
 	"devDependencies": {
-		"ava": "*",
-		"xo": "*"
+		"ava": "^3.15.0",
+		"xo": "^0.46.3"
 	}
 }
