@@ -1,27 +1,27 @@
 import test from 'ava';
-import m from '.';
+import darkMode from './index.js';
 
 let isDarkInitially;
 
 test.before(async () => {
-	isDarkInitially = await m.isDark();
+	isDarkInitially = await darkMode.isDark();
 });
 
 test.after(async () => {
-	await m.toggle(isDarkInitially);
+	await darkMode.toggle(isDarkInitially);
 });
 
-test(async t => {
-	const dark = await m.isDark();
-	await m.toggle();
-	t.not(dark, await m.isDark());
+test('main', async t => {
+	const dark = await darkMode.isDark();
+	await darkMode.toggle();
+	t.not(dark, await darkMode.isDark());
 
-	await m.disable();
-	t.false(await m.isDark());
+	await darkMode.disable();
+	t.false(await darkMode.isDark());
 
-	await m.enable();
-	t.true(await m.isDark());
+	await darkMode.enable();
+	t.true(await darkMode.isDark());
 
-	await m.toggle(false);
-	t.false(await m.isDark());
+	await darkMode.toggle(false);
+	t.false(await darkMode.isDark());
 });
